Changed SCHD message to use a variable and int to store server type and id, and fixed bug
SCHD message has been changed from using an array containing servers and their ids to simply using a string with the largest server type and an incrementing int for the server id.

Also fixed a bug that occurred near the start when initiating contact with the server due to OK messages being sent consecutively and not clearing the input string, resulting in the client sending messages before the server is ready.